requirements.txt
fastapi==0.135.1
httpx==0.28.1
imageio-ffmpeg==0.6.0
openai-whisper==20250625
python-multipart==0.0.22
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
uvicorn==0.42.0
